Reorder the MADCTL bits to match the datasheet
When none of the MV/X/Y bits is set, ie. no transformation, the
top-left of the display is such that the ribbon cable of the LCD
module enters the top edge, with column numbers increasing from
left-to-right, and row numbers increasing upon moving further
down the panel, away from the ribbon cable.

This PR corrects the declaration of the MV, MX and MY bits and
updates the orientation settings accordingly so client code using
the API should be unaffected.